👷 add --no-delete flag to deploy scripts

+++ deploy.sh
@@ -4,16 +4,28 @@ BUILD_DIR="dist"
 PI_USER="admin"
 PI_HOST="192.168.178.43"
 PI_TARGET_DIR="/home/admin/rpi-led-matrix-application/webapp"
+DELETE_OLD_FILES=true
+
+for arg in "$@"
+do
+    if [[ $arg == "--no-delete" ]]; then
+        DELETE_OLD_FILES=false
+        break
+    fi
+done
 
 echo "Building the project..."
 npm run build || { echo "Build failed"; exit 1; }
 
 echo
 
-echo "Deleting old files on Raspberry Pi..."
-ssh $PI_USER@$PI_HOST "rm -rf $PI_TARGET_DIR/*" || { echo "Failed to delete old files"; exit 1; }
-
-echo "Done."
+if [ "$DELETE_OLD_FILES" = true ]; then
+    echo "Deleting old files on Raspberry Pi..."
+    ssh $PI_USER@$PI_HOST "rm -rf $PI_TARGET_DIR/*" || { echo "Failed to delete old files"; exit 1; }
+    echo "Done."
+else
+    echo "Skipping deletion of old files."
+fi
 
 echo
